[REFACTOR] #100 - Dialog -> DialogFragment로 구현 수정
